requirements.txt
reflex==0.3.10
openai~=1.11.1
python-dotenv~=1.0.1
